System working, merging with main

+++ BlazorWASM-AppLogic/HttpClients/ClientInterfaces/IPostService.cs
@@ -10,7 +10,7 @@ public interface IPostService
 {
     Task<int> CreateAsync(PostCreationDto postCreationDto);
     Task<ICollection<Post>> GetAsync(int? id = null, int? userId = null, string? titleContains = null);
-    Task DeleteAsync(int id, long user_id);
+    Task DeleteAsync(long id, long user_id);
     Task UpdateAsync(UpdatePostDto dto, long user_id);
     Task<IEnumerable<Post>> GetInRadiusAsync(Coordinate coordinate, int radius);
 }
