+++ BlazorWASM-AppLogic/HttpClients/Implementations/PostHttpClient.cs
@@ -100,7 +100,7 @@ public class PostHttpClient : IPostService
 	    return query;
     }
 
-    public async Task DeleteAsync(int id, long user_id)
+    public async Task DeleteAsync(long id, long user_id)
     {
 
 	    HttpResponseMessage response = await client.DeleteAsync($"https://localhost:7196/posts/{id}?user_id={user_id}");
